requirements.txt
requests~=2.32.5
pandas~=2.3.2
yadisk~=3.4.0
Flask~=3.1.2
python-dotenv~=1.1.1
python-calamine~=0.4.0
gunicorn~=22.0.0